+++ client/packages/system/src/index.ts
@@ -17,3 +17,4 @@ export * from './Log';
 export * from './ReturnReason';
 export * from './Currency';
 export * from './IndicatorsDemographics';
+export { processProperties } from './utils';
